+++ src/core_modules/capture-core/components/Pages/common/TEIRelationshipsWidget/TeiSearch/actions/teiSearch.actions.js
@@ -3,21 +3,22 @@
 import { actionCreator } from '../../../../../../actions/actions.utils';
 
 export const batchActionTypes = {
-    BATCH_SET_TEI_SEARCH_PROGRAM_AND_TET: 'BatchSetTeiSearchProgramAndTet',
-    RESET_SEARCH_FORMS: 'ResetSearchForms',
+    BATCH_SET_TEI_SEARCH_PROGRAM_AND_TET: 'RelationshipsWidget.BatchSetTeiSearchProgramAndTet',
+    RESET_SEARCH_FORMS: 'RelationshipsWidget.ResetSearchForms',
 };
 
 export const actionTypes = {
-    INITIALIZE_TEI_SEARCH: 'InitializeTeiSearch',
-    REQUEST_SEARCH_TEI: 'RequestSearchTei',
-    SEARCH_FORM_VALIDATION_FAILED: 'SearchFormValidationFailed',
-    SEARCH_TEI_FAILED: 'SearchTeiFailed',
-    SEARCH_TEI_RESULT_RETRIEVED: 'SearchTeiResultRetrieved',
-    SET_TEI_SEARCH_PROGRAM_AND_TET: 'SetTeiSearchProgramAndTet',
-    TEI_NEW_SEARCH: 'TeiNewSearch',
-    TEI_EDIT_SEARCH: 'TeiEditSearch',
-    TEI_SEARCH_RESULTS_CHANGE_PAGE: 'TeiSearchResultsChangePage',
-    TEI_SEARCH_SET_OPEN_SEARCH_GROUP_SECTION: 'TeiSearchSetOpenSearchGroupSection',
+    INITIALIZE_TEI_SEARCH: 'RelationshipsWidget.InitializeTeiSearch',
+    REQUEST_SEARCH_TEI: 'RelationshipsWidget.RequestSearchTei',
+    SEARCH_FORM_VALIDATION_FAILED: 'RelationshipsWidget.SearchFormValidationFailed',
+    SEARCH_TEI_FAILED: 'RelationshipsWidget.SearchTeiFailed',
+    SEARCH_TEI_RESULT_RETRIEVED: 'RelationshipsWidget.SearchTeiResultRetrieved',
+    SET_TEI_SEARCH_PROGRAM_AND_TET: 'RelationshipsWidget.SetTeiSearchProgramAndTet',
+    TEI_NEW_SEARCH: 'RelationshipsWidget.TeiNewSearch',
+    TEI_EDIT_SEARCH: 'RelationshipsWidget.TeiEditSearch',
+    TEI_SEARCH_RESULTS_CHANGE_PAGE: 'RelationshipsWidget.TeiSearchResultsChangePage',
+    TEI_SEARCH_SET_OPEN_SEARCH_GROUP_SECTION: 'RelationshipsWidget.TeiSearchSetOpenSearchGroupSection',
+    SEARCH_TE_IN_TET_SCOPE: 'RelationshipsWidget.SearchTrackedEntityInTETScope',
 };
 
 
@@ -31,6 +32,27 @@ export const requestSearchTei = (
     resultsPageSize: number,
 ) =>
     actionCreator(actionTypes.REQUEST_SEARCH_TEI)({ formId, searchGroupId, searchId, resultsPageSize });
+
+export const searchViaUniqueIdOnScopeTrackedEntityType = ({
+    formId,
+    searchGroupId,
+    searchId,
+    selectedProgramId,
+    programQueryArgs,
+}: {
+    formId: string,
+    searchGroupId: string,
+    searchId: string,
+    selectedProgramId: string,
+    programQueryArgs: any,
+}) =>
+    actionCreator(actionTypes.SEARCH_TE_IN_TET_SCOPE)({
+        formId,
+        searchGroupId,
+        searchId,
+        selectedProgramId,
+        programQueryArgs,
+    });
 
 export const searchTeiFailed = (
     formId: string,