+++ src/core_modules/capture-core/components/SearchBox/SearchForm/SearchForm.epics.js
@@ -12,6 +12,7 @@ import {
     addSuccessResultsViewOnSearchBox,
     showTooManyResultsViewOnSearchBox,
     showFallbackNotEnoughAttributesOnSearchBox,
+    searchViaUniqueIdOnScopeTrackedEntityType,
 } from '../SearchBox.actions';
 import {
     getTrackedEntityInstances,
@@ -42,23 +43,33 @@ const searchViaUniqueIdStream = ({
     programId,
     absoluteApiPath,
     querySingleResource,
+    formId,
+    programTETId,
 }: {
     queryArgs: any,
     attributes: any,
     programId?: string,
     absoluteApiPath: string,
     querySingleResource: QuerySingleResource,
+    formId?: string,
+    programTETId?: string,
 }) =>
-    from(getTrackedEntityInstances(queryArgs, attributes, absoluteApiPath, querySingleResource)).pipe(
+    from(getTrackedEntityInstances(queryArgs, attributes, absoluteApiPath, querySingleResource, programId)).pipe(
         flatMap(({ trackedEntityInstanceContainers }) => {
             const searchResults = trackedEntityInstanceContainers;
+            if (searchResults.length === 0 && queryArgs.program) {
+                return of(searchViaUniqueIdOnScopeTrackedEntityType({ trackedEntityTypeId: programTETId, formId }));
+            }
             if (searchResults.length > 0) {
-                const { id, tei: { orgUnit: orgUnitId } } = searchResults[0];
+                const { id, tei: { orgUnit: orgUnitId, enrollments } } = searchResults[0];
+                const programToNavigateTo = enrollments?.length === 1 && !programId
+                    ? enrollments[0].program
+                    : programId;
 
                 return of(navigateToEnrollmentOverview({
                     teiId: id,
                     orgUnitId,
-                    programId,
+                    programId: programToNavigateTo,
                 }));
             }
             return of(showEmptyResultsViewOnSearchBox());
@@ -92,8 +103,22 @@ const handleErrors = ({ httpStatusCode, message }) => {
     return of(showErrorViewOnSearchBox());
 };
 
-const searchViaAttributesStream = ({ queryArgs, attributes, triggeredFrom, absoluteApiPath, querySingleResource }) =>
-    from(getTrackedEntityInstances(queryArgs, attributes, absoluteApiPath, querySingleResource)).pipe(
+const searchViaAttributesStream = ({
+    queryArgs,
+    attributes,
+    triggeredFrom,
+    absoluteApiPath,
+    querySingleResource,
+    programId,
+}: {
+    queryArgs: any,
+    attributes: any,
+    triggeredFrom: string,
+    absoluteApiPath: string,
+    querySingleResource: QuerySingleResource,
+    programId?: string,
+}) =>
+    from(getTrackedEntityInstances(queryArgs, attributes, absoluteApiPath, querySingleResource, programId)).pipe(
         map(({ trackedEntityInstanceContainers: searchResults, pagingData }) => {
             if (searchResults.length > 0) {
                 return showSuccessResultsViewOnSearchBox(
@@ -136,7 +161,7 @@ export const searchViaUniqueIdOnScopeProgramEpic = (
                 ouMode: 'ACCESSIBLE',
             };
 
-            const attributes = getTrackerProgramThrowIfNotFound(programId).attributes;
+            const { attributes, trackedEntityType } = getTrackerProgramThrowIfNotFound(programId);
 
             return searchViaUniqueIdStream({
                 queryArgs,
@@ -144,6 +169,8 @@ export const searchViaUniqueIdOnScopeProgramEpic = (
                 programId,
                 absoluteApiPath,
                 querySingleResource,
+                formId,
+                programTETId: trackedEntityType.id,
             });
         }),
     );
@@ -165,6 +192,7 @@ export const searchViaUniqueIdOnScopeTrackedEntityTypeEpic = (
                 trackedEntityType: trackedEntityTypeId,
                 pageNumber: 1,
                 ouMode: 'ACCESSIBLE',
+                fields: 'trackedEntity,trackedEntityType,orgUnit,attributes,enrollments',
             };
 
             const attributes = getTrackedEntityTypeThrowIfNotFound(trackedEntityTypeId).attributes;
@@ -204,6 +232,7 @@ export const searchViaAttributesOnScopeProgramEpic = (
                 triggeredFrom,
                 absoluteApiPath,
                 querySingleResource,
+                programId,
             });
         }),
     );